Split captured-variables mode logic into its own file
pause_point_wait.go grew past the 500-line architecture guard
(TestProductionGoFilesStayFocused) once the --captured-variables
mode type, parser, and stripping logic were added. Moved that
functionally-cohesive block (mode type, parse, apply, strip helper)
into pause_point_captured_variables_mode.go, no behavior change.
